Save keyword and paper metadata to database
Previously, metadata (keyword name, paper title, paper abstarct, etc.)
would be stored in a .pickle file, along with the embeddings. This behavior
has been udpated to store the metadata in a SQL database for more straightforward
SQL commands to get top-scoring paper data from the query.

This also makes the module more flexible to allow for source paper and keyword data
to come in database format.