Separate initial database storage of file data
This is to make the module more flexible, since data can be provided in both
file and database form. If provided in file form, store_to_database.py should be
run to store the data in a database. Then, store_embeddings.py can be run on the
data in the database