Update method of mapping papers to embeddings
Upon storing paper embeddings (in store_embeddings.py) and additional
file mapping paper ID's to their index in the embedding matrix is stored.
Before, it was assumed that the paper's ID would match the index of the
embedding, however this assumption could break under various circumstances.

As of now, the paper ID is the corresponding ID of the paper in the source file,
but this could be changed to account for future circumstances.

This assumpetion is still met for keywords, since our keyword source file doesn't
include keyword IDs