Improve performance by removing unnecessary joins
get_ranked_publications was joining with the Publication table
to return Publication titles. This, however, caused significant slowdowns
because the LEFT function was being used before the GROUP BY, causing joins
on rows that we're going to be aggregated. Doing the JOIN after the
GROUP BY reduces the number of searches MySQL has to do, causining significant
performance improvements (approximate reduction from 7s to 2s). For now, the
JOIN has been removed entirely, and the function returns only the Publication id
and the match score