Improve performance with executemany
Originaly, assign_paper_keywords would call execute() individually on
every row, using a for loop, which was inefficient. Instead, we can
store all the data required for storage in a list and add all the data
to the database at once using executemany(), which is significantly faster

While testing with 10000 papers, the script took approximately 25s before the change.
After the change, the script takes approximately 11s